[refactor] formatDate utils 분리
